# generated/java/src/main/java/saas/identity/shared/api/MeApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.24.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package saas.identity.shared.api;

import saas.identity.shared.dto.CurrentUser;
import saas.identity.shared.dto.ErrorResponse;
import saas.identity.shared.dto.SwitchTenantResponse;
import saas.identity.shared.dto.TenantMembership;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-12T12:50:18.473810800+08:00[Asia/Shanghai]", comments = "Generator version: 7.24.0")
@Validated
@Tag(name = "me", description = "the me API")
public interface MeApi {

    String PATH_ME_LIST_MY_TENANTS = "/api/v1/me/tenants";
    /**
     * GET /api/v1/me/tenants
     *
     * @return The request has succeeded. (status code 200)
     *         or An unexpected error response. (status code 200)
     */
    @Operation(
        operationId = "meListMyTenants",
        tags = { "me" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request has succeeded.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TenantMembership.class)))
            }),
            @ApiResponse(responseCode = "default", description = "An unexpected error response.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = MeApi.PATH_ME_LIST_MY_TENANTS,
        produces = { "application/json" }
    )
    ResponseEntity<List<TenantMembership>> meListMyTenants(
        
    );


    String PATH_ME_SWITCH_TENANT = "/api/v1/me/tenants/{tenantId}/switch";
    /**
     * POST /api/v1/me/tenants/{tenantId}/switch
     *
     * @param tenantId  (required)
     * @return The request has succeeded. (status code 200)
     *         or An unexpected error response. (status code 200)
     */
    @Operation(
        operationId = "meSwitchTenant",
        tags = { "me" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request has succeeded.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SwitchTenantResponse.class))
            }),
            @ApiResponse(responseCode = "default", description = "An unexpected error response.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = MeApi.PATH_ME_SWITCH_TENANT,
        produces = { "application/json" }
    )
    ResponseEntity<SwitchTenantResponse> meSwitchTenant(
        @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") String tenantId
    );


    String PATH_ME_WHOAMI = "/api/v1/me";
    /**
     * GET /api/v1/me
     *
     * @return The request has succeeded. (status code 200)
     *         or An unexpected error response. (status code 200)
     */
    @Operation(
        operationId = "meWhoami",
        tags = { "me" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request has succeeded.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CurrentUser.class))
            }),
            @ApiResponse(responseCode = "default", description = "An unexpected error response.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = MeApi.PATH_ME_WHOAMI,
        produces = { "application/json" }
    )
    ResponseEntity<CurrentUser> meWhoami(
        
    );

}