# generated/java/src/main/java/saas/identity/shared/api/TenantApiKeysApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.24.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package saas.identity.shared.api;

import saas.identity.shared.dto.ApiKey;
import saas.identity.shared.dto.CreateApiKeyRequest;
import saas.identity.shared.dto.CreateApiKeyResponse;
import saas.identity.shared.dto.ErrorResponse;
import org.springframework.lang.Nullable;
import saas.identity.shared.dto.TenantApiKeysListApiKeys200Response;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-12T12:50:18.473810800+08:00[Asia/Shanghai]", comments = "Generator version: 7.24.0")
@Validated
@Tag(name = "tenant-api-keys", description = "the tenant-api-keys API")
public interface TenantApiKeysApi {

    String PATH_TENANT_API_KEYS_CREATE_API_KEY = "/api/v1/tenants/{tenantId}/api-keys";
    /**
     * POST /api/v1/tenants/{tenantId}/api-keys
     *
     * @param tenantId  (required)
     * @param createApiKeyRequest  (required)
     * @return The request has succeeded. (status code 200)
     *         or An unexpected error response. (status code 200)
     */
    @Operation(
        operationId = "tenantApiKeysCreateApiKey",
        tags = { "tenant-api-keys" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request has succeeded.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreateApiKeyResponse.class))
            }),
            @ApiResponse(responseCode = "default", description = "An unexpected error response.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = TenantApiKeysApi.PATH_TENANT_API_KEYS_CREATE_API_KEY,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<CreateApiKeyResponse> tenantApiKeysCreateApiKey(
        @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") String tenantId,
        @Parameter(name = "CreateApiKeyRequest", description = "", required = true) @Valid @RequestBody CreateApiKeyRequest createApiKeyRequest
    );


    String PATH_TENANT_API_KEYS_LIST_API_KEYS = "/api/v1/tenants/{tenantId}/api-keys";
    /**
     * GET /api/v1/tenants/{tenantId}/api-keys
     *
     * @param tenantId  (required)
     * @param page  (optional)
     * @param pageSize  (optional)
     * @return The request has succeeded. (status code 200)
     *         or An unexpected error response. (status code 200)
     */
    @Operation(
        operationId = "tenantApiKeysListApiKeys",
        tags = { "tenant-api-keys" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request has succeeded.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TenantApiKeysListApiKeys200Response.class))
            }),
            @ApiResponse(responseCode = "default", description = "An unexpected error response.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TenantApiKeysApi.PATH_TENANT_API_KEYS_LIST_API_KEYS,
        produces = { "application/json" }
    )
    ResponseEntity<TenantApiKeysListApiKeys200Response> tenantApiKeysListApiKeys(
        @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") String tenantId,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false) @Nullable Integer page,
        @Parameter(name = "pageSize", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageSize", required = false) @Nullable Integer pageSize
    );


    String PATH_TENANT_API_KEYS_REVOKE_API_KEY = "/api/v1/tenants/{tenantId}/api-keys/{keyId}/revoke";
    /**
     * POST /api/v1/tenants/{tenantId}/api-keys/{keyId}/revoke
     *
     * @param tenantId  (required)
     * @param keyId  (required)
     * @return The request has succeeded. (status code 200)
     *         or An unexpected error response. (status code 200)
     */
    @Operation(
        operationId = "tenantApiKeysRevokeApiKey",
        tags = { "tenant-api-keys" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request has succeeded.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiKey.class))
            }),
            @ApiResponse(responseCode = "default", description = "An unexpected error response.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = TenantApiKeysApi.PATH_TENANT_API_KEYS_REVOKE_API_KEY,
        produces = { "application/json" }
    )
    ResponseEntity<ApiKey> tenantApiKeysRevokeApiKey(
        @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") String tenantId,
        @Parameter(name = "keyId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("keyId") String keyId
    );


    String PATH_TENANT_API_KEYS_ROTATE_API_KEY = "/api/v1/tenants/{tenantId}/api-keys/{keyId}/rotate";
    /**
     * POST /api/v1/tenants/{tenantId}/api-keys/{keyId}/rotate
     *
     * @param tenantId  (required)
     * @param keyId  (required)
     * @return The request has succeeded. (status code 200)
     *         or An unexpected error response. (status code 200)
     */
    @Operation(
        operationId = "tenantApiKeysRotateApiKey",
        tags = { "tenant-api-keys" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request has succeeded.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreateApiKeyResponse.class))
            }),
            @ApiResponse(responseCode = "default", description = "An unexpected error response.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = TenantApiKeysApi.PATH_TENANT_API_KEYS_ROTATE_API_KEY,
        produces = { "application/json" }
    )
    ResponseEntity<CreateApiKeyResponse> tenantApiKeysRotateApiKey(
        @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") String tenantId,
        @Parameter(name = "keyId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("keyId") String keyId
    );

}