# generated/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  version: 0.0.0
...
paths:
...
  /api/v1/me:
    get:
      operationId: Me_whoami
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - me
  /api/v1/me/tenants:
    get:
      operationId: Me_listMyTenants
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TenantMembership'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - me
  /api/v1/me/tenants/{tenantId}/switch:
    post:
      operationId: Me_switchTenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwitchTenantResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - me
  /api/v1/tenants/{tenantId}/api-keys:
    get:
      operationId: TenantApiKeys_listApiKeys
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - page
                  - pageSize
                  - total
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
                  page:
                    type: integer
                    format: int32
                    example: 0
                  pageSize:
                    type: integer
                    format: int32
                    example: 20
                  total:
                    type: integer
                    format: int64
                    example: 0
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-api-keys
    post:
      operationId: TenantApiKeys_createApiKey
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApiKeyResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-api-keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
  /api/v1/tenants/{tenantId}/api-keys/{keyId}/revoke:
    post:
      operationId: TenantApiKeys_revokeApiKey
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-api-keys
  /api/v1/tenants/{tenantId}/api-keys/{keyId}/rotate:
    post:
      operationId: TenantApiKeys_rotateApiKey
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApiKeyResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-api-keys
...
components:
  schemas:
    ApiKey:
      type: object
      required:
        - id
        - tenantId
        - name
        - prefix
        - status
        - scopes
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 2
          maxLength: 128
        prefix:
          type: string
          minLength: 8
          maxLength: 16
        status:
          $ref: '#/components/schemas/ApiKeyStatus'
        scopes:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
    ApiKeyStatus:
      type: string
      enum:
        - active
        - revoked
        - expired
...
    CreateApiKeyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 128
        scopes:
          type: array
          items:
            type: string
        expiresAt:
          type: string
          format: date-time
    CreateApiKeyResponse:
      type: object
      required:
        - apiKey
        - secret
      properties:
        apiKey:
          $ref: '#/components/schemas/ApiKey'
        secret:
          type: string
          minLength: 16
          maxLength: 256
...
    CurrentUser:
      type: object
      required:
        - id
        - email
        - memberships
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        memberships:
          type: array
          items:
            $ref: '#/components/schemas/TenantMembership'
        currentTenantId:
          type: string
          format: uuid
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
...
    MembershipStatus:
      type: string
      enum:
        - active
        - invited
        - removed
...
    SwitchTenantResponse:
      type: object
      required:
        - accessToken
        - expiresAt
        - tenantId
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresAt:
          type: string
          format: date-time
        tenantId:
          type: string
          format: uuid
...
    TenantMembership:
      type: object
      required:
        - id
        - userId
        - tenantId
        - roleIds
        - status
        - joinedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        roleIds:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/MembershipStatus'
        joinedAt:
          type: string
          format: date-time
...
servers:
  - url: https://api.example.com
    description: Production
    variables: {}